New view listing all checkpoints
This view is going to be developed into an admin feature. Template
name is likely to be changed or the feature moved onto the
existing checkpoint.html template.

New function users.list_checkpoints(): returns all checkpoints in
the database as a dict where user_id is a key, other entries values.